feat: Add user preferences file system with persistence
Add the ability for the user to set preferences for the UI and save them to a json file. Also implement a persistence feature where the user's preferences can be updated and persist beyond closing and opening a new window. Include small bug fixes and refactoring to certain components and functions.